+++ HISTORY.rst
@@ -25,8 +25,8 @@ History
 .. Security in case of vulnerabilities.
 
 
-Unreleased
-~~~~~~~~~~
+18.11.0 (2018-11-09)
+~~~~~~~~~~~~~~~~~~~~
 
 Added
 ^^^^^